--- lib/action.ts
@@ -1,1 +1,0 @@
-'use server';
